+++ crates/amaru-uplc/src/machine/cost_model/mod.rs
@@ -26,13 +26,13 @@ pub mod cost_map;
 pub mod costing;
 pub mod ex_budget;
 
+pub mod cost_argument;
+
 mod param_name;
 pub use param_name::*;
 
 mod step_kind;
 pub use step_kind::*;
-
-pub mod value;
 
 #[derive(Debug, PartialEq, Default)]
 pub struct CostModel {